# api/openapi/tms-public/openapi.yaml
openapi: 3.1.0
info:
  title: Hyperlocalise TMS Public API
  version: 0.1.0
  summary: Public integration surface for Hyperlocalise TMS.
servers:
  - url: https://api.hyperlocalise.dev
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/jobs:
    get:
      tags: [Jobs]
      operationId: listJobs
      summary: List jobs
      responses:
        "200":
          description: A list of jobs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
    post:
      tags: [Jobs]
      operationId: createJob
      summary: Submit a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "202":
          description: Job accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, kind, status]
      properties:
        id:
          type: string
        kind:
          type: string
        status:
          type: string
        projectId:
          type: string
        createdAt:
          type: string
          format: date-time
...
    CreateJobRequest:
      type: object
      required: [projectId, kind]
      properties:
        projectId:
          type: string
        kind:
          type: string
        resourceIds:
          type: array
          items:
            type: string
...
    JobListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Job"
...